restructure and combine OCPs in parent class
