+++ dynamics/__init__.py
@@ -1,0 +1,2 @@
+from .dynamics_centroidal import DynamicsCentroidal
+from .dynamics_rnea import DynamicsRNEA
