+++ optimal_control_problem/__init__.py
@@ -1,0 +1,3 @@
+from .ocp import OCP
+from .ocp_centroidal import OCP_Centroidal
+from .ocp_rnea import OCP_RNEA
